EIP1-3666 Refactored directory name
